+++ WeatherIs.OpenWeatherMapApi/Models/IndividualModels/OneCallApi/Minutely.cs
@@ -4,7 +4,7 @@ namespace WeatherIs.OpenWeatherMapApi.Models.IndividualModels.OneCallApi
 {
     public class Minutely
     {
-        [JsonProperty("dt")] public int Dt { get; set; }
+        [JsonProperty("dt")] public long Dt { get; set; }
 
         [JsonProperty("precipitation")] public double Precipitation { get; set; }
     }
